fix: correct tests
