+++ __tests__/tests/LayoutPage.test.tsx
@@ -3,11 +3,10 @@ import '@testing-library/jest-dom';
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { BrowserRouter } from 'react-router-dom';
-import { useAuthState } from 'react-firebase-hooks/auth';
 import LayoutPage from '../../src/pages/LayoutPage/LayoutPage';
 
 jest.mock('react-firebase-hooks/auth', () => ({
-  useAuthState: jest.fn() as jest.Mock,
+  useAuthState: jest.fn().mockReturnValue([null, false, null]),
 }));
 
 describe('Test LayoutPage', () => {
@@ -17,7 +16,6 @@ describe('Test LayoutPage', () => {
         <LayoutPage />
       </BrowserRouter>
     );
-    (useAuthState as jest.Mock).mockReturnValue([null, false, null]);
     const signInBtn = screen.getByText(/Sign In/i);
     const user = userEvent.setup();
     user.click(signInBtn);
@@ -31,31 +29,11 @@ describe('Test LayoutPage', () => {
         <LayoutPage />
       </BrowserRouter>
     );
-    (useAuthState as jest.Mock).mockReturnValue([null, false, null]);
     const signUpBtn = screen.getByText(/Sign Up/i);
     const user = userEvent.setup();
     user.click(signUpBtn);
     await waitFor(() => {
       expect(window.location.pathname).toEqual('/registration');
     });
-  });
-  it('Checking if the user is authorized there is a button Sign Out ', async () => {
-    render(
-      <BrowserRouter>
-        <LayoutPage />
-      </BrowserRouter>
-    );
-    (useAuthState as jest.Mock).mockReturnValue([
-      {
-        user: {
-          displayName: 'John Doe',
-          email: 'johndoe@example.com',
-        },
-        loading: false,
-        error: null,
-      },
-    ]);
-    const signOut = screen.getByText(/Sign Out/i);
-    expect(signOut).toBeInTheDocument();
   });
 });
